Move encode_call from evm_transaction to evm_helpers
encode_call is a contract-call encoder, not part of the transaction send
flow — relocate it to evm_helpers alongside the other EVM helpers. All 22
import sites rewired (the shared `import encode_call, send_evm_transaction`
lines split across evm_helpers/evm_transaction).
