+++ wayfinder_paths/adapters/ethena_vault_adapter/adapter.py
@@ -17,7 +17,8 @@ from wayfinder_paths.core.constants.ethena_contracts import (
     ethena_tokens_by_chain_id,
 )
 from wayfinder_paths.core.utils.evm_client import web3_from_chain_id
-from wayfinder_paths.core.utils.evm_transaction import encode_call, send_evm_transaction
+from wayfinder_paths.core.utils.evm_helpers import encode_call
+from wayfinder_paths.core.utils.evm_transaction import send_evm_transaction
 from wayfinder_paths.core.utils.interest import apr_to_apy
 from wayfinder_paths.core.utils.multicall import (
     Call,
