+++ wayfinder_paths/adapters/etherfi_adapter/adapter.py
@@ -22,7 +22,8 @@ from wayfinder_paths.core.constants.etherfi_contracts import (
     ETHERFI_BY_CHAIN,
 )
 from wayfinder_paths.core.utils.evm_client import web3_from_chain_id
-from wayfinder_paths.core.utils.evm_transaction import encode_call, send_evm_transaction
+from wayfinder_paths.core.utils.evm_helpers import encode_call
+from wayfinder_paths.core.utils.evm_transaction import send_evm_transaction
 from wayfinder_paths.core.utils.multicall import (
     Call,
     read_only_calls_multicall_or_gather,
