+++ wayfinder_paths/adapters/euler_v2_adapter/adapter.py
@@ -20,7 +20,8 @@ from wayfinder_paths.core.constants.euler_v2_abi import (
 )
 from wayfinder_paths.core.constants.euler_v2_contracts import EULER_V2_BY_CHAIN
 from wayfinder_paths.core.utils.evm_client import web3_from_chain_id
-from wayfinder_paths.core.utils.evm_transaction import encode_call, send_evm_transaction
+from wayfinder_paths.core.utils.evm_helpers import encode_call
+from wayfinder_paths.core.utils.evm_transaction import send_evm_transaction
 from wayfinder_paths.core.utils.interest import RAY
 from wayfinder_paths.core.utils.tokens import ensure_allowance
