+++ wayfinder_paths/adapters/projectx_adapter/adapter.py
@@ -29,8 +29,8 @@ from wayfinder_paths.core.constants.uniswap_v3_abi import (
     NONFUNGIBLE_POSITION_MANAGER_ABI,
 )
 from wayfinder_paths.core.utils.evm_client import web3_from_chain_id
+from wayfinder_paths.core.utils.evm_helpers import encode_call
 from wayfinder_paths.core.utils.evm_transaction import (
-    encode_call,
     send_evm_transaction,
     wait_for_evm_transaction,
 )
