+++ wayfinder_paths/adapters/uniswap_adapter/base.py
@@ -10,7 +10,8 @@ from wayfinder_paths.core.constants.uniswap_v3_abi import (
     UNISWAP_V3_FACTORY_ABI,
 )
 from wayfinder_paths.core.utils.evm_client import web3_from_chain_id
-from wayfinder_paths.core.utils.evm_transaction import encode_call, send_evm_transaction
+from wayfinder_paths.core.utils.evm_helpers import encode_call
+from wayfinder_paths.core.utils.evm_transaction import send_evm_transaction
 from wayfinder_paths.core.utils.tokens import ensure_allowance
 from wayfinder_paths.core.utils.uniswap_v3_math import (
     PositionData,
