+++ wayfinder_paths/core/utils/tokens.py
@@ -14,7 +14,8 @@ from wayfinder_paths.core.constants.erc20_abi import (
 )
 from wayfinder_paths.core.constants.erc1155_abi import ERC1155_APPROVAL_ABI
 from wayfinder_paths.core.utils.evm_client import web3_from_chain_id
-from wayfinder_paths.core.utils.evm_transaction import encode_call, send_evm_transaction
+from wayfinder_paths.core.utils.evm_helpers import encode_call
+from wayfinder_paths.core.utils.evm_transaction import send_evm_transaction
 
 NATIVE_TOKEN_ADDRESSES: set = {
     "0x0000000000000000000000000000000000000000",
