+++ wayfinder_paths/strategies/boros_hype_strategy/withdraw_mixin.py
@@ -18,7 +18,8 @@ from wayfinder_paths.adapters.hyperliquid_adapter.paired_filler import (
     PairedFiller,
 )
 from wayfinder_paths.core.strategies.Strategy import StatusTuple
-from wayfinder_paths.core.utils.evm_transaction import encode_call, send_evm_transaction
+from wayfinder_paths.core.utils.evm_helpers import encode_call
+from wayfinder_paths.core.utils.evm_transaction import send_evm_transaction
 
 from .constants import (
     ARBITRUM_CHAIN_ID,
